refactor: SIMT-45 user usecase 패키지 정리
