+++ domain/src/main/java/com/comit/domain/usecase/users/SignUpUseCase.kt
@@ -1,23 +1,31 @@
-package com.comit.domain.usecase
+package com.comit.domain.usecase.users
 
 import com.comit.domain.repository.AuthRepository
 import javax.inject.Inject
 
-class SignInUseCase @Inject constructor(
+class SignUpUseCase @Inject constructor(
     private val repository: AuthRepository,
 ) {
 
     suspend operator fun invoke(
         params: Params,
     ) = kotlin.runCatching {
-        repository.signIn(
+        repository.signUp(
+            name = params.name,
             employeeNumber = params.employeeNumber,
+            email = params.email,
             password = params.password,
+            nickname = params.nickname,
+            profileImagePath = params.profileImagePath,
         )
     }
 
     data class Params(
+        val name: String,
         val employeeNumber: Int,
+        val email: String,
         val password: String,
+        val nickname: String?,
+        val profileImagePath: String?,
     )
 }
